Add PlayView with updateField method closes #11
Creates a new window in which the game will be played in. The class has
a method updateScene to draw the field on a scene and update the view
to show whats drawn.

Each SquareFigure is programed so that if it’s clicked it calls the
click(x, y) method with the x & y values of the clicked Square in the
field and then calls the updateField to show the new scene.